Cancel method integration test (#240)
* Migrate progress test to upscale

* Add integration test for cancel
