+++ .gitignore
@@ -20,3 +20,4 @@ upscaled.png
 scratch
 packages/upscalerjs/src/tfjs.ts
 *.generated.ts
+.vscode
